Vendor npcs: removing redundant StoreItem and StoreItemReconstitutionFactory
